+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## [1.1.0]
 ### Changed
-- New supported IDE version range: 2024.2 - 2024.3.*.
+- New supported IDE version range: 2024.2.1 - 2024.3.*.
 - Updated the project to use the IntelliJ Platform Gradle Plugin 2.0.
 - Updated the project to use JDK 21.
 - Updated project configuration to make sure the plugin works when the K2 Kotlin compiler is enabled.